change order deletion form; fix bugs

+++ Controller/src/main/webapp/client/js/order/adminDeleteOrder.js
@@ -21,7 +21,13 @@ function getOrderDetailedInfo(order_id) {
             $("#transactionCount").find('span').html(data.transactionCount);
             $("#companyCommission").find('span').html(data.companyCommission?data.companyCommission + ' ' + data.currencyConvertName:'-');
             /**/
-            $("#delete-order-info__delete").attr('onclick', 'deleteOrderByAdmin(' + data.id + ')');
+            if (data.orderStatusName.toUpperCase() === 'DELETED') {
+                $("#delete-order-info__delete").toggle(false);
+            } else {
+                $("#delete-order-info__delete").toggle(true);
+                $("#delete-order-info__delete").attr('onclick', 'deleteOrderByAdmin(' + data.id + ')');
+            }
+
             /**/
             $('#order-delete-modal').modal();
         }
@@ -66,44 +72,43 @@ function searchAndDeleteOrderByAdmin() {
 function searchOrder() {
     var isError = false;
     $('.input-block-wrapper__error-wrapper').toggle(false);
-    var match = $('#orderRate').val().match(/^\d+(\.{1,1}\d+)?/);
-    if (!match || match[0] !== $('#orderRate').val()){
-        $('.input-block-wrapper__error-wrapper[for=orderRate]').toggle(true);
+    var orderRate = $('#orderRate').val();
+    var orderVolume = $('#orderVolume').val();
+    var creatorEmail = $('#creatorEmail').val();
+
+    if (orderRate.length > 0) {
+        var match = orderRate.match(/^\d+(\.{1,1}\d+)?/);
+        if (!match || match[0] !== orderRate){
+            $('.input-block-wrapper__error-wrapper[for=orderRate]').toggle(true);
+            isError = true;
+        }
+    }
+    if (orderVolume.length > 0) {
+        match = orderVolume.match(/^\d+(\.{1,1}\d+)?/);
+        if (!match || match[0] !== orderVolume){
+            $('.input-block-wrapper__error-wrapper[for=orderVolume]').toggle(true);
+            isError = true;
+        }
+    }
+
+    if (creatorEmail.length > 0) {
+        match = creatorEmail.match(/^[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Za-z]{2,}$/);
+        if (!match || match[0] !== creatorEmail){
+            $('.input-block-wrapper__error-wrapper[for=creatorEmail]').toggle(true);
+            isError = true;
+        }
+    }
+
+    if ($('#orderDateFrom').val().length > 0 && (!$('#orderDateFrom').val().match(/\d{4}\-\d{2}\-\d{2}\s{1}\d{2}\:\d{2}\:\d{2}/))){
+        $('.input-block-wrapper__error-wrapper[for=orderDateFrom]').toggle(true);
         isError = true;
     }
-    match = $('#orderVolume').val().match(/^\d+(\.{1,1}\d+)?/);
-    if (!match || match[0] !== $('#orderVolume').val()){
-        $('.input-block-wrapper__error-wrapper[for=orderVolume]').toggle(true);
-        isError = true;
-    }
-    if (!$('#orderDate').val().match(/\d{4}\-\d{2}\-\d{2}\s{1}\d{2}\:\d{2}\:\d{2}/)){
-        $('.input-block-wrapper__error-wrapper[for=orderDate]').toggle(true);
+    if ($('#orderDateTo').val().length > 0 && !($('#orderDateTo').val().match(/\d{4}\-\d{2}\-\d{2}\s{1}\d{2}\:\d{2}\:\d{2}/))){
+        $('.input-block-wrapper__error-wrapper[for=orderDateTo]').toggle(true);
         isError = true;
     }
     if (isError) {
         return;
     }
-    var data = $('#delete-order-info__form').serialize();
-    $.ajax({
-        url: '/admin/searchorder',
-        type: 'GET',
-        data: data,
-        success: function (data) {
-            $('#order-delete-modal--result-info').find('.delete-order-info__item').toggle(false);
-            if (data === -1) {
-                $('#order-delete-modal--result-info').find('.error-search').toggle(true);
-                $('#order-delete-modal--result-info').modal();
-            } else {
-                $('#order-delete-modal--search').one('hidden.bs.modal', function (e) {
-                    getOrderDetailedInfo(data);
-                });
-                $('#order-delete-modal--search').modal('hide');
-            }
-        },
-        error: function(){
-            $('#order-delete-modal--result-info').find('.delete-order-info__item').toggle(false);
-            $('#order-delete-modal--result-info').find('.error-delete').toggle(true);
-            $('#order-delete-modal--result-info').modal();
-        }
-    });
+    $('#delete-order-info__form').submit();
 }
